Move some barcode types to there own classes
