+++ src/Types/Type.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Picqer\Barcode\Types;
+
+abstract class Type
+{
+    abstract public function getBarcodeData(string $code): array;
+}
